backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg[binary]==3.3.2
python-dotenv==1.0.0
argon2-cffi==23.1.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0